+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Add `ignoredCount` as templating variable. ([#191])
 
 ## [0.4.4] - 2020-04-12
 
@@ -125,3 +125,4 @@ Versioning].
 [#185]: https://github.com/ericcornelissen/svgo-action/pull/185
 [#186]: https://github.com/ericcornelissen/svgo-action/pull/186
 [#187]: https://github.com/ericcornelissen/svgo-action/pull/187
+[#191]: https://github.com/ericcornelissen/svgo-action/pull/191